remove redundant schemas from model_connectors
